mcp/fill-simu-mcp/pyproject.toml
[project]
name = "fill-simu-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.3",
    "pyyaml>=6.0.3",
]
